agents/01-01-hello-gemini/requirements.txt
langchain
langchain-openai
openai
google-generativeai
python-dotenv